+++ ts/features/it-wallet/screens/presentation/crossdevice/new/ItwPresentationResultScreen.tsx
@@ -4,7 +4,7 @@ import { useNavigation } from "@react-navigation/native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { IOStyles } from "@pagopa/io-app-design-system";
 import { itwPresentation } from "../../../../store/actions/new/itwPresentationActions";
-import { itwPresentationChecksSelector } from "../../../../store/reducers/new/itwPresentationReducer";
+import { itwPresentationResultSelector } from "../../../../store/reducers/new/itwPresentationReducer";
 import { useIODispatch, useIOSelector } from "../../../../../../store/hooks";
 import { useOnFirstRender } from "../../../../../../utils/hooks/useOnFirstRender";
 import ItwLoadingSpinnerOverlay from "../../../../components/ItwLoadingSpinnerOverlay";
@@ -27,7 +27,7 @@ import { getItwGenericMappedError } from "../../../../utils/errors/itwErrorsMapp
  */
 const ItwPresentationResultScreen = () => {
   const dispatch = useIODispatch();
-  const checksPot = useIOSelector(itwPresentationChecksSelector);
+  const resultPot = useIOSelector(itwPresentationResultSelector);
   const navigation =
     useNavigation<IOStackNavigationProp<ItwParamsList & AppParamsList>>();
   const rpMock = getRpMock();
@@ -39,7 +39,10 @@ const ItwPresentationResultScreen = () => {
   const LoadingView = () => (
     <ItwLoadingSpinnerOverlay
       captionTitle={I18n.t(
-        "features.itWallet.presentation.checksScreen.loading"
+        "features.itWallet.presentation.resultScreen.loading.title",
+        {
+          organizationName: rpMock.organizationName
+        }
       )}
       isLoading
     >
@@ -57,10 +60,10 @@ const ItwPresentationResultScreen = () => {
     <SafeAreaView style={IOStyles.flex}>
       <ItwKoView
         title={I18n.t(
-          "features.itWallet.presentation.resultScreenNew.success.title"
+          "features.itWallet.presentation.resultScreen.success.title"
         )}
         subtitle={I18n.t(
-          "features.itWallet.presentation.resultScreenNew.success.subtitle",
+          "features.itWallet.presentation.resultScreen.success.subtitle",
           {
             organizationName: rpMock.organizationName
           }
@@ -78,7 +81,7 @@ const ItwPresentationResultScreen = () => {
 
   const RenderMask = () =>
     pot.fold(
-      checksPot,
+      resultPot,
       () => <LoadingView />,
       () => <LoadingView />,
       () => <LoadingView />,